Revert "#49 Split blocks unit tests (#38)"
This reverts commit f994672b6346b416370468a22982bc4819f25673.
