+++ wagtail_wordpress_import/block_builder.py
@@ -32,12 +32,9 @@ IMAGE_SRC_DOMAIN = "https://www.budgetsaresexy.com"  # note no trailing /
 
 
 class BlockBuilder:
-    def __init__(self, value, node, logger):
+    def __init__(self, value):
         self.soup = BeautifulSoup(value, "lxml", exclude_encodings=True)
         self.blocks = []
-        self.logged_items = {"processed": 0, "imported": 0, "skipped": 0, "items": []}
-        self.node = node
-        self.logger = logger
         self.set_up()
 
     def set_up(self):
@@ -68,12 +65,12 @@ class BlockBuilder:
         for tag in soup:
             counter += 1
             """
-            the process here loops though each soup tag to discover
-            the block type to use
+            the process here loops though each soup tag to discover the block type to use
+            there's a table and iframe and form block to deal with if they exist
             """
 
             # RICHTEXT
-            if tag.name not in TAGS_TO_BLOCKS:
+            if not tag.name in TAGS_TO_BLOCKS:
                 block_value += str(self.image_linker(str(tag)))
 
             # TABLE
@@ -84,6 +81,10 @@ class BlockBuilder:
                 self.blocks.append({"type": "raw_html", "value": str(tag)})
 
             # IFRAME/EMBED
+            """
+            test escaped code to add to xml for parsing
+            &lt;iframe width=&quot;560&quot; height=&quot;315&quot; src=&quot;https://www.youtube.com/embed/CQ7Gx8b7ac4&quot; title=&quot;YouTube video player&quot; frameborder=&quot;0&quot; allow=&quot;accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture&quot; allowfullscreen&gt;&lt;/iframe&gt;
+            """
             if tag.name == "iframe":
                 if len(block_value) > 0:
                     self.blocks.append({"type": "rich_text", "value": block_value})
@@ -98,6 +99,16 @@ class BlockBuilder:
                 )
 
             # FORM
+
+            """
+            test escaped code to add to xml for parsing
+            &lt;form&gt;
+            &lt;label for=&quot;fname&quot;&gt;First name:&lt;/label&gt;&lt;br&gt;
+            &lt;input type=&quot;text&quot; id=&quot;fname&quot; name=&quot;fname&quot;&gt;&lt;br&gt;
+            &lt;label for=&quot;lname&quot;&gt;Last name:&lt;/label&gt;&lt;br&gt;
+            &lt;input type=&quot;text&quot; id=&quot;lname&quot; name=&quot;lname&quot;&gt;
+            &lt;/form&gt;
+            """
             if tag.name == "form":
                 if len(block_value) > 0:
                     self.blocks.append({"type": "rich_text", "value": block_value})
@@ -151,7 +162,6 @@ class BlockBuilder:
                 self.blocks.append({"type": "rich_text", "value": block_value})
                 block_value = str("")
 
-        # print(self.logged_items)
         return self.blocks
 
     def image_linker(self, tag):
@@ -187,22 +197,7 @@ class BlockBuilder:
             temp = NamedTemporaryFile(delete=True)
             image_src = check_image_src(image.get("src")).strip("/")
         else:
-            self.logged_items["items"].append(
-                {
-                    "id": self.node.get("wp:post_id"),
-                    "title": self.node.get("title"),
-                    "link": self.node.get("link"),
-                    "reason": "no src provided",
-                }
-            )
-            self.logger.images.append(
-                {
-                    "id": self.node.get("wp:post_id"),
-                    "title": self.node.get("title"),
-                    "link": self.node.get("link"),
-                    "reason": "no src provided",
-                }
-            )
+            print("WARNING: found an src with no value")
             return
 
         try:
@@ -213,32 +208,18 @@ class BlockBuilder:
 
             try:
                 response = requests.get(image_src, timeout=10)
-                status_code = response.status_code
+                status = response.status_code
                 content_type = response.headers.get("Content-Type")
 
-                if (
-                    content_type
-                    and content_type.lower() not in VALID_IMAGE_CONTENT_TYPES
-                ):
-                    self.logged_items["items"].append(
-                        {
-                            "id": self.node.get("wp:post_id"),
-                            "title": self.node.get("title"),
-                            "link": self.node.get("link"),
-                            "reason": "invalid image types match or no content type",
-                        }
-                    )
-                    self.logger.images.append(
-                        {
-                            "id": self.node.get("wp:post_id"),
-                            "title": self.node.get("title"),
-                            "link": self.node.get("link"),
-                            "reason": "invalid image types match or no content type",
-                        }
+                if not content_type in VALID_IMAGE_CONTENT_TYPES:
+                    print(
+                        "WARNING: No content type returned or invalid content type for image {}. This is what we know {}".format(
+                            image_src, content_type
+                        )
                     )
                     return
 
-                if status_code == 200:
+                if status == 200 and content_type.lower() in VALID_IMAGE_CONTENT_TYPES:
                     temp.name = name
                     temp.write(response.content)
                     temp.flush()
@@ -247,21 +228,10 @@ class BlockBuilder:
                     return new_image
 
             except requests.exceptions.ConnectionError:
-                self.logged_items["items"].append(
-                    {
-                        "id": self.node.get("wp:post_id"),
-                        "title": self.node.get("title"),
-                        "link": self.node.get("link"),
-                        "reason": "connection error",
-                    }
-                )
-                self.logger.images.append(
-                    {
-                        "id": self.node.get("wp:post_id"),
-                        "title": self.node.get("title"),
-                        "link": self.node.get("link"),
-                        "reason": "connection error",
-                    }
+                print(
+                    "WARNING: Unable to connect to URL {}. Image will be broken.".format(
+                        image.get("src")
+                    )
                 )
