+++ wagtail_wordpress_import/importers/wordpress.py
@@ -1,27 +1,28 @@
 import json
+import sys
 from datetime import datetime
 from functools import cached_property
 from xml.dom import pulldom
 
 from django.apps import apps
-from django.conf import settings
-from django.utils.module_loading import import_string
 from django.utils.text import slugify
 from django.utils.timezone import make_aware
 from wagtail.core.models import Page
+from wagtail_wordpress_import.bleach import bleach_clean, fix_styles
 from wagtail_wordpress_import.block_builder import BlockBuilder
-from wagtail_wordpress_import.functions import node_to_dict
-from wagtail_wordpress_import.prefilters.linebreaks_wp_filter import (
-    filter_linebreaks_wp,
+from wagtail_wordpress_import.functions import (
+    linebreaks_wp,
+    node_to_dict,
+    normalize_style_attrs,
 )
 
 
 class WordpressImporter:
     def __init__(self, xml_file_path):
         self.xml_file = xml_file_path
+        self.logged_items = {"processed": 0, "imported": 0, "skipped": 0, "items": []}
 
     def run(self, *args, **kwargs):
-        logger = kwargs["logger"]
         xml_doc = pulldom.parse(self.xml_file)
 
         try:
@@ -49,14 +50,14 @@ class WordpressImporter:
             if event == pulldom.START_ELEMENT and node.tagName == "item":
                 xml_doc.expandNode(node)
                 item = node_to_dict(node)
-                logger.processed += 1
+                self.logged_items["processed"] += 1
 
                 if (
                     item.get("wp:post_type") in kwargs["page_types"]
                     and item.get("wp:status") in kwargs["page_statuses"]
                 ):
 
-                    wordpress_item = WordpressItem(item, logger)
+                    wordpress_item = WordpressItem(item)
 
                     try:
                         page = self.page_model_instance.objects.get(
@@ -74,8 +75,7 @@ class WordpressImporter:
 
                     if page.id:
                         page.save()
-                        logger.imported += 1
-                        logger.items.append(
+                        self.logged_items["items"].append(
                             {
                                 "id": page.id,
                                 "title": page.title,
@@ -86,11 +86,12 @@ class WordpressImporter:
                                 "slugcheck": wordpress_item.slug_changed,
                             }
                         )
+                        self.logged_items["imported"] += 1
+
                     else:
                         self.parent_page_obj.add_child(instance=page)
                         page.save()
-                        logger.imported += 1
-                        logger.items.append(
+                        self.logged_items["items"].append(
                             {
                                 "id": page.id,
                                 "title": page.title,
@@ -101,9 +102,9 @@ class WordpressImporter:
                                 "slugcheck": wordpress_item.slug_changed,
                             }
                         )
+                        self.logged_items["imported"] += 1
                 else:
-                    logger.skipped += 1
-                    logger.items.append(
+                    self.logged_items["items"].append(
                         {
                             "id": 0,
                             "title": "",
@@ -114,8 +115,9 @@ class WordpressImporter:
                             "slugcheck": "",
                         }
                     )
+                    self.logged_items["skipped"] += 1
             else:
-                logger.items.append(
+                self.logged_items["items"].append(
                     {
                         "id": 0,
                         "title": "",
@@ -126,7 +128,17 @@ class WordpressImporter:
                         "slugcheck": "",
                     }
                 )
-            logger.log_progress()
+            self.log_to_console(self.logged_items["items"][-1])
+
+        return self.logged_items
+
+    def log_to_console(self, item):
+        if not item["id"] == 0:
+            sys.stdout.write(
+                f"{item['id']}, {item['title']}, {item['result']}, {self.logged_items['processed']}\n"
+            )
+        else:
+            sys.stdout.write(f"skipped ... {self.logged_items['processed']}\n")
 
     def analyze_html(self, html_analyzer, *, page_types, page_statuses):
         xml_doc = pulldom.parse(self.xml_file)
@@ -144,60 +156,30 @@ class WordpressImporter:
                     stream_fields = self.mapping_stream_fields.split(",")
 
                     for html in stream_fields:
-                        value = filter_linebreaks_wp(
+                        value = linebreaks_wp(
                             item.get(self.mapping_item_inverse.get(html))
                         )
                         html_analyzer.analyze(value)
 
 
-DEFAULT_PREFILTERS = [
-    {
-        "FUNCTION": "wagtail_wordpress_import.prefilters.linebreaks_wp_filter.filter_linebreaks_wp",
-    },
-    {
-        "FUNCTION": "wagtail_wordpress_import.prefilters.normalize_styles_filter.filter_normalize_style_attrs",
-    },
-    {
-        "FUNCTION": "wagtail_wordpress_import.prefilters.fix_styles_filter.filter_fix_styles",
-    },
-    {
-        "FUNCTION": "wagtail_wordpress_import.prefilters.bleach_filter.filter_bleach_clean",
-    },
-]
-
-DEBUG_ENABLED = getattr(settings, "WAGTAIL_WORDPRESS_IMPORT_DEBUG", True)
-
-
 class WordpressItem:
-    def __init__(self, node, logger):
+    def __init__(self, node):
         self.node = node
         self.raw_body = self.node["content:encoded"]
+        self.linebreaks_wp = linebreaks_wp(
+            self.raw_body
+        )  # generally adds a p tag where it finds a linebreak
+        self.normalize = normalize_style_attrs(
+            self.linebreaks_wp
+        )  # formats style attrs to be lower cased and correctly spaced with trailing ; on each style
+        self.fix_styles = fix_styles(
+            self.normalize
+        )  # takes a complete style attr and alters the html to reflect the style required
+        self.bleach_clean = bleach_clean(
+            self.fix_styles
+        )  # stanity check to remove illegal/iincorrect html
         self.slug_changed = ""
         self.date_changed = ""
-        self.image_errors = []
-        # self.url_errors = []
-
-        self.debug_content = {}
-        self.logger = logger
-
-    def prefilter_content(self, content):
-        """
-        FILTERS ARE CUMULATIVE
-        cache the result of each filter which is run on the output from the previous filter
-        """
-        filter_config = getattr(
-            settings, "WAGTAIL_WORDPRESS_IMPORT_PREFILTERS", DEFAULT_PREFILTERS
-        )
-
-        cached_result = content
-
-        for filter in filter_config:
-            function = import_string(filter["FUNCTION"])
-            cached_result = function(cached_result, filter.get("OPTIONS"))
-            if DEBUG_ENABLED:
-                self.debug_content[function.__name__] = cached_result
-
-        return cached_result
 
     def cleaned_title(self):
         return self.node["title"].strip()
@@ -256,11 +238,8 @@ class WordpressItem:
     def cleaned_link(self):
         return str(self.node["link"].strip())
 
-    def body_stream_field(self, content):
-        blocks_dict = BlockBuilder(content, self.node, self.logger).build()
-        if DEBUG_ENABLED:
-            self.debug_content["block_json"] = blocks_dict
-        return json.dumps(blocks_dict)
+    def cleaned_body(self):
+        return json.dumps(BlockBuilder(self.bleach_clean).build())
 
     @cached_property
     def cleaned_data(self):
@@ -270,14 +249,12 @@ class WordpressItem:
             "first_published_at": self.cleaned_first_published_at(),
             "last_published_at": self.cleaned_last_published_at(),
             "latest_revision_created_at": self.cleaned_latest_revision_created_at(),
-            "body": self.body_stream_field(self.prefilter_content(self.raw_body)),
+            "body": self.cleaned_body(),
             "wp_post_id": self.cleaned_post_id(),
             "wp_post_type": self.cleaned_post_type(),
             "wp_link": self.cleaned_link(),
-            "wp_block_json": self.debug_content.get("block_json"),
-            "wp_processed_content": self.debug_content.get("filter_fix_styles"),
-            "wp_normalized_styles": self.debug_content.get(
-                "filter_normalize_style_attrs"
-            ),
-            "wp_raw_content": self.debug_content.get("filter_linebreaks_wp"),
+            "wp_raw_content": self.raw_body,
+            "wp_processed_content": self.fix_styles,
+            "wp_block_json": self.cleaned_body(),
+            "wp_normalized_styles": self.normalize,
         }
