--- wagtail_wordpress_import/prefilters/__init__.py
@@ -1,3 +1,0 @@
-default_app_config = (
-    "wagtail_wordpress_import.prefilters.apps.WagtailWordpressImportPrefiltersConfig"
-)
