--- wagtail_wordpress_import/prefilters/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class WagtailWordpressImportPrefiltersConfig(AppConfig):
-    label = "wagtail_wordpress_import_prefilters"
-    name = "wagtail_wordpress_import_prefilters"
-    verbose_name = "Wagtail wordpress import prefilters"
