+++ wagtail_wordpress_import/test/tests/test_wordpress_item.py
@@ -3,7 +3,6 @@ import json
 from django.test import TestCase
 from datetime import datetime
 from wagtail_wordpress_import.importers.wordpress import WordpressItem
-from wagtail_wordpress_import.logger import Logger
 
 BASE_PATH = os.path.dirname(os.path.dirname(__file__))
 FIXTURES_PATH = BASE_PATH + "/fixtures"
@@ -11,7 +10,7 @@ FIXTURES_PATH = BASE_PATH + "/fixtures"
 
 class WordpressItemTests(TestCase):
     def setUp(self):
-        self.logger = Logger("fakedir")
+        # logging.disable(logging.CRITICAL)
         raw_html_file = open(f"{FIXTURES_PATH}/raw_html.txt", "r").read()
         self.good_node = {
             "title": "Page Title",
@@ -35,7 +34,7 @@ class WordpressItemTests(TestCase):
         }
 
     def test_all_fields_with_good_data(self):
-        wordpress_item = WordpressItem(self.good_node, self.logger)
+        wordpress_item = WordpressItem(self.good_node)
         title = wordpress_item.cleaned_data["title"]
         slug = wordpress_item.cleaned_data["slug"]
         first_published_at = wordpress_item.cleaned_data["first_published_at"]
@@ -47,9 +46,9 @@ class WordpressItemTests(TestCase):
         wp_post_id = wordpress_item.cleaned_data["wp_post_id"]
         wp_post_type = wordpress_item.cleaned_data["wp_post_type"]
         wp_link = wordpress_item.cleaned_data["wp_link"]
-        wp_raw_content = wordpress_item.debug_content["filter_linebreaks_wp"]
-        wp_processed_content = wordpress_item.debug_content["filter_fix_styles"]
-        wp_block_json = wordpress_item.debug_content["block_json"]
+        wp_raw_content = wordpress_item.cleaned_data["wp_raw_content"]
+        wp_processed_content = wordpress_item.cleaned_data["wp_processed_content"]
+        wp_block_json = wordpress_item.cleaned_data["wp_block_json"]
 
         self.assertEqual(title, "Page Title")
         self.assertEqual(slug, "page-title")
@@ -62,10 +61,10 @@ class WordpressItemTests(TestCase):
         self.assertEqual(wp_link, "http://www.example.com")
         self.assertIsInstance(wp_raw_content, str)
         self.assertIsInstance(wp_processed_content, str)
-        self.assertIsInstance(wp_block_json, list)
+        self.assertIsInstance(wp_block_json, str)
 
     def test_cleaned_fields(self):
-        wordpress_item = WordpressItem(self.bad_node, self.logger)
+        wordpress_item = WordpressItem(self.bad_node)
         slug = wordpress_item.cleaned_data["slug"]
         first_published_at = wordpress_item.cleaned_data["first_published_at"]
         last_published_at = wordpress_item.cleaned_data["last_published_at"]
